Create atlas_brain_matching(16_bit_output)
I've added a function to the end of this script called atlasBrainMatch16_bit which finds a folder of pngs in the same folder as the directory of pngs used for the DLC landmark prediction and applies the transformation to that. The altered funciton will maintain 16 depth through to the output images. You can use this to apply transformations from landmarks found on 8 bit images from totally different images, or perhaps another channel used in hemodynamic correction. This can be useful if the fluorescent image has poorer or uneven contrast across the image. We had this issue due to virus injections that didn't evenly distribute across the brain.